+++ ratticweb/static/rattic/css/base.css
@@ -1,17 +1,13 @@
-.parent-sized {
-  width: 100%;
-}
-
 .no-display {
   display: none;
 }
 
 .center {
-    text-align:center:
+    text-align:center;
 }
 
 .col-lg-10 h1 {
-    margin-left:-15px:
+    margin-left:-15px;
 }
 
 div.copybutton {
