+++ ratticweb/templates/base.html
@@ -24,7 +24,7 @@
   <link href="{% static "bootstrap/css/bootstrap.min.css" %}" rel="stylesheet" media="screen">
   <link href="{% static "bootstrap/css/bootstrap-theme.min.css" %}" rel="stylesheet" media="screen">
   <link href="{% static "selectize/0.12.1/css/selectize.css" %}" rel="stylesheet" media="screen">
-  <link href="{% static "selectize/0.21.1/css/selectize.bootstrap3.css" %}" rel="stylesheet" media="screen">
+  <link href="{% static "selectize/0.12.1/css/selectize.bootstrap3.css" %}" rel="stylesheet" media="screen">
   <link href="{% static "rattic/css/base.css" %}" rel="stylesheet" media="screen">
   <link href="{% static "rattic/css/icons.css" %}" rel="stylesheet" media="screen">
   <script src="{% static "modernizr/2.6.2/modernizr.js" %}" ></script>
@@ -54,7 +54,7 @@
 </div>
 <script type="application/javascript" src="{% static "jquery/js/jquery.min.js" %}"></script>
 <script type="application/javascript" src="{% static "bootstrap/js/bootstrap.min.js" %}"></script>
-<script type="application/javascript" src="{% static "selectize/0.12.1/js/selectize.min.js" %}"></script>
+<script type="application/javascript" src="{% static "selectize/0.12.1/js/selectize.js" %}"></script>
 <script type="application/javascript" src="{% static "zeroclipboard/1.3.2/ZeroClipboard.min.js" %}"></script>
 <script type="application/javascript" src="{% static "flash_detect/1.0.4/flash_detect_min.js" %}"></script>
 <script type="application/javascript" src="{% static "sjcl/1.0.0/sjcl.js" %}"></script>
