+++ ratticweb/templates/home.html
@@ -9,7 +9,7 @@
     {% blocktrans %}Your username and password didn't match. Please try again.{% endblocktrans %}
 </div>
 {% endif %}
-<div class="pagination-centered">
+<div class="center">
     <h1>{% trans "Welcome to Rattic" %}</h1>
 	<img class="biglogo" alt="{% trans "RatticDB Logo" %}" src="{% static rattic_logo %}">
     <p>{% trans "It does passwords." %}</p>
