+++ ratticweb/templates/navbar.html
@@ -11,7 +11,7 @@
     <div class="navbar-header">
         <a class="navbar-brand" href="{% url_root %}"><img class="logo" src="{% static rattic_icon %}" /><span class="hidden-xs"> {% trans "RatticDB" %}</span></a>
     </div>
-    <nav class="nav  navbar-default">
+    <nav class="nav navbar-default">
 
         <ul class="nav navbar-nav visible-md visible-lg">
             <li {% if pageurl == cred_views_list %}class="active"{% endif %}>
