+++ ratticweb/templates/sidebar.html
@@ -5,7 +5,7 @@
 {% if user.is_authenticated %}
     <h4>{% trans "You" %}</h4>
         <p>{% blocktrans with username=user.username|title %}Welcome {{ username }}.{% endblocktrans %}<p>
-        <a class="btn" href="{% url "account.views.profile" %}">{% trans "Profile" %}</a>
+        <a class="btn btn-default" href="{% url "account.views.profile" %}">{% trans "Profile" %}</a>
         <a class="btn btn-danger" href="{% url "django.contrib.auth.views.logout" %}">{% trans "Logout" %}</a>
     <h4><a href="{% url "cred.views.tags"  %}">{% trans "Tags" %}</a></h4>
     {% if user.is_authenticated %}
@@ -25,9 +25,9 @@
 {% else %}
 <h4>{% trans "Login" %}</h4>
     <form class="form-inline" method="post" action="{% url "login" %}?next={{ next|default:'' }}">{% csrf_token %}
-			<label for="id_username">{% trans "Username" %}</label><input class="parent-sized" id="id_username" type="text" name="auth-username" maxlength="30" />
-			<label for="id_password">{% trans "Password" %}</label><input class="parent-sized" type="password" name="auth-password" id="id_password" />
-			<input type="submit" class="btn" value="{% trans "Login" %}"/>
+			<label for="id_username">{% trans "Username" %}</label><input class="form-control" id="id_username" type="text" name="auth-username" maxlength="30" />
+			<label for="id_password">{% trans "Password" %}</label><input class="form-control" type="password" name="auth-password" id="id_password" />
+			<input type="submit" class="btn btn-default" value="{% trans "Login" %}"/>
 	{% if not LDAP_ENABLED %}<p><a href="{% url "django.contrib.auth.views.password_reset" %}">{% trans "Forgotten password?" %}</a></p>{% endif %}
     <input type="hidden" name="rattic_tfa_login_view-current_step" value="auth" />
     </form>
